+++ R/anova_apa.R
@@ -17,8 +17,9 @@
 #'   (generalized eta squared) (you may also use the abbreviations \code{"pes"}
 #'   or \code{"ges"}).
 #' @param format Character string specifying the output format. One of
-#'   \code{"text"}, \code{"markdown"}, \code{"rmarkdown"}, \code{html},
-#'   \code{"latex"}, \code{"latex_math"}, \code{"docx"} or \code{"plotmath"}.
+#'   \code{"text"}, \code{"markdown"}, \code{"quarto"}, \code{"rmarkdown"},
+#'   \code{html}, \code{"latex"}, \code{"latex_math"}, \code{"docx"} or
+#'   \code{"plotmath"}.
 #' @param info Logical indicating whether to print a message on the used test
 #'   (default is \code{FALSE})
 #' @param print Logical indicating whether to print the formatted output via
@@ -48,8 +49,9 @@ anova_apa <- function(x, effect = NULL,
                                    "hf", "none"),
                       force_sph_corr = FALSE,
                       es = c("petasq", "pes", "getasq", "ges"),
-                      format = c("text", "markdown", "rmarkdown", "html",
-                                 "latex", "latex_math", "docx", "plotmath"),
+                      format = c("text", "markdown", "quarto", "rmarkdown",
+                                 "html", "latex", "latex_math", "docx",
+                                 "plotmath"),
                       info = FALSE, print = TRUE)
 {
   sph_corr <- match.arg(sph_corr)