+++ R/apa.R
@@ -1,7 +1,7 @@
-#' APA Formatting for RMarkdown Reports
+#' APA Formatting for Quarto and RMarkdown Reports
 #'
 #' A wrapper around the \code{*_apa} functions, providing a convenient way to
-#' use the formatters in inline code in RMarkdown documents.
+#' use the formatters in inline code in Quarto and RMarkdown documents.
 #'
 #' @param x An \R object. Must be a call to one of \code{afex::aov_4},
 #'   \code{afex::aov_car}, \code{afex::aov_ez}, \code{chisq.test},
@@ -9,8 +9,8 @@
 #' @param effect (only applicable if \code{x} is an ANOVA) Character string
 #'   indicating the name of the effect to display.
 #' @param format Character string specifying the output format. One of
-#'   \code{"text"}, \code{"markdown"}, \code{"rmarkdown"}, \code{html},
-#'   \code{"latex"} or \code{"docx"}.
+#'   \code{"text"}, \code{"markdown"}, \code{"quarto"}, \code{"rmarkdown"},
+#'   \code{html}, \code{"latex"} or \code{"docx"}.
 #' @param print Logical indicating whether to return the result as an \R object
 #'   (\code{FALSE}) or print using \code{cat} (\code{TRUE}).
 #' @param ... Further arguments passed to other methods
@@ -24,15 +24,15 @@ apa <- function(x, effect = NULL, format = "rmarkdown", print = FALSE, ...)
   {
     if (grepl("Chi-squared test", x$method))
     {
-      chisq_apa(x, format = format, print = print, ...)
+      out <- chisq_apa(x, format = format, print = print, ...)
     }
     else if (grepl("correlation", x$method))
     {
-      cor_apa(x, format = format, print = print, ...)
+      out <- cor_apa(x, format = format, print = print, ...)
     }
     else if (grepl("t-test", x$method))
     {
-      t_apa(x, format = format, print = print, ...)
+      out <- t_apa(x, format = format, print = print, ...)
     }
     else
     {
@@ -47,10 +47,13 @@ apa <- function(x, effect = NULL, format = "rmarkdown", print = FALSE, ...)
       stop("For calls to ANOVA, an `effect` must be specified.")
     }
 
-    anova_apa(x, effect, format = format, print = print, ...)
+    out <- anova_apa(x, effect, format = format, print = print, ...)
   }
   else
   {
     stop("Unkown type passed to 'x'")
   }
+
+  # Prevent escaping of markdown and math syntax in Quarto rendering
+  I(out)
 }