+++ R/cor_apa.R
@@ -5,8 +5,9 @@
 #'   for the correlation coefficient (default is \code{FALSE}). Only available
 #'   for Pearson's product moment correlation (with n >= 4).
 #' @param format Character string specifying the output format. One of
-#'   \code{"text"}, \code{"markdown"}, \code{"rmarkdown"}, \code{html},
-#'   \code{"latex"}, \code{"latex_math"}, \code{"docx"} or \code{"plotmath"}.
+#'   \code{"text"}, \code{"markdown"}, \code{"quarto"}, \code{"rmarkdown"},
+#'   \code{html}, \code{"latex"}, \code{"latex_math"}, \code{"docx"} or
+#'   \code{"plotmath"}.
 #' @param info Logical indicating whether to print a message on the used test
 #'   (default is \code{FALSE})
 #' @param print Logical indicating whether to print the formatted output via
@@ -26,8 +27,9 @@
 #'
 #' @export
 cor_apa <- function(x, r_ci = FALSE,
-                    format = c("text", "markdown", "rmarkdown", "html", "latex",
-                               "latex_math", "docx", "plotmath"),
+                    format = c("text", "markdown", "quarto", "rmarkdown",
+                               "html", "latex", "latex_math", "docx",
+                               "plotmath"),
                     info = FALSE, print = TRUE)
 {
   format <- match.arg(format)